Create README.md
My mission with the need to stimulate high intelli-qvote is to hack electric gear, join me and teach and give pointer and ideas for me to succeed. I learn instantly with the need of the same kind of minds that understands the importance of clarity of information. Let us have some fun!